feat : Poolservice Kafka

+++ CRM-BE/StatisticService/src/main/java/org/bitznomad/statisticservice/repository/StatisticRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatisticRepo extends JpaRepository<Statistic, Integer> {
+public interface StatisticRepo extends JpaRepository<Statistic, Long> {
 
 }
